Update the method to check the intersection (line-plane intersection)
Previous, Jarred's method was not efficient, he took the distance of each vertex of the Cat.ply with a point of the light curtain, but it was not the right method that we should do. I did change to Line Plane intersection collision check method. In this code, I draw a line from the paw of the cat to the tail, this will be the line and the plane is the light curtain plane. Then I move the line and cat.ply simultaneously until the line touch the plane.